Fixed a bug when adding entry and now the user can add float entries

+++ database/migrations/2025_12_21_141219_create_foods_table.php
@@ -14,23 +14,22 @@ return new class extends Migration
             $table->string('brand')->nullable();
             $table->string('name');
             $table->string('name_hu')->nullable();
-            
-            $table->integer('calorie')->default(0);
-            $table->integer('protein')->default(0);
-            $table->integer('fat')->default(0);
-            $table->integer('carb')->default(0);
+
+            $table->decimal('calorie', 8, 2)->default(0);
+            $table->decimal('protein', 8, 2)->default(0);
+            $table->decimal('fat', 8, 2)->default(0);
+            $table->decimal('carb', 8, 2)->default(0);
 
             $table->string('unit')->default('g');
-            $table->integer('amount')->default(100);
-            
+            $table->decimal('amount', 8, 2)->default(100);
+
             $table->text('image')->nullable();
-            
+
             $table->timestamps();
-            
+
             $table->index(['name', 'name_hu', 'brand']);
         });
     }
-
     public function down(): void
     {
         Schema::dropIfExists('foods');
